Change ui to light mode and refactor comparators
